bulk_insert() in the documentation has: obj = (     MyModel.objects     .on_conflict(['name'], ConflictAction.UPDATE)     .bulk_insert([ ...
However bulk_insert() wasn't actually returning anything. What would be useful would be to utilize the 'returning' clause already being added to the Postgres query to send back any auto inc indexes the table might have. This makes the functionality better mirror bulk_create()

In addition, when using the on_conflict() this would allow the user to see the indexes for any entry that might be updated.

As well, to make this work there was a bug in how 'returning' clause rows were being retrieved in the compiler, it was only grabbing the first entry from the cursor, not all the returned records.

Tests have been added to this patch as well as documentation.