Have `PostgresQuerySet.insert` always return the primary key
This fixes #48 and ensures that objects which specify an explicit
primary key value will have that value returned from 'insert'.

This makes `insert` reliably usable for models which don't have
an `id` field (which previously `AttributeError`ed or returned
`None`, depending on whether a conflict was anticipated).

Note: this is potentially quietly breaking for models which:
- specified a non-`id` primary key, and
- also had a separate `id` field
This scenario seems sufficiently unlikely to be worth the fix.