Use get_compiler instead of compiler to avoid breaking users
Setting the `compiler` attribute allows the db back-end to actually
look up the compiler. This is useful because the back-end can
actually return different compilers depending on the query/config.

However, this is problematic if the back-end is not the one you
expect it to be. One of our internal products has this problem
and this recent change breaks their set up. Since this release
already has quite a few breaking changes. Let's do away with this one.