Changed how annotations are temporarily renamed
If only a subset of annotations is renamed before calling the base class,
and then renamed back using add new key/remove old key, then the ordering of
annotations is changed (annotations are stored in a OrderedDict internally).
This will cause issues if ordering is important (e.g., in a union).