Attempt at fixing #158 by returning raw values from the DB
`return_model=True` would break any rows with JSONField as it
would try to convert from a JSON string to a Python object. However,
the original input data was already partially converted.

By forcing the database to return the values it inserted, we
get a raw JSON string and the converted applies normally.