Inspect base database back-end instance to figure out base classes
The PostGIS back-end doesn't set `DatabaseWrapper.ops_class` properly
and instead initializes its own `Operations` class in the constructor.

This results in django-postgres-extra subclassing the wrong base class.

As a solution, we create an instance and inspect what class is actually
used for `DatabaseWrapper.ops_class` and `DatabaseWrapper.introspection_class`
and subclass from that.

[#181799346]