Fix an issue whether detection of update/insert fields isn't reliable
This is because `pre_save` mutates the model instance. Since we call
it twice (once with `is_insert=True` and once with `is_insert=False`),
the second call might have a different result.

This problem manifests with `DateTimeField(auto_now=True)`. It'll always
be treated as an insert field, even though it it should be treated
as both.